Scratch from the website login sistem
The PHP script 'connect.php' makes a connection with the data base, and
should be included in every script that needs it. The password is a md5
hash on the database and is associated with a login.

In the begining of the login page 'index.php' is done a verification of
the login and password, if those variables where given. If not the page
is showed normally.